Re-built map in preparation for merge with location pulling
