--- ui/src/domain/domain.ts
@@ -1,9 +1,0 @@
-export interface HotChocolate {
-    productName: string,
-    englishProductName: string,
-    vendor: string,
-    location: string,
-    lat: number,
-    lon: number,
-    description?: string
-};
